perf: make improvements

+++ P05/ppos_data.h
@@ -13,13 +13,13 @@
 typedef struct task_t
 {
   struct task_t *prev, *next ;		// ponteiros para usar em filas
-  int id ;				// identificador da tarefa
+  unsigned int id ;				// identificador da tarefa
   ucontext_t context ;			// contexto armazenado da tarefa
   short status ;			// pronta, rodando, suspensa, ...
-  int static_prio ;			// prioridade estática da tarefa
-  int dynamic_prio ;			// prioridade dinâmica da tarefa
+  short int static_prio ;			// prioridade estática da tarefa
+  short int dynamic_prio ;			// prioridade dinâmica da tarefa
   short type ;				// tipo de tarefa (user, system)
-  unsigned int quantum ;			// ticks restantes para a tarefa
+  short unsigned int quantum ;			// ticks restantes para a tarefa
   // ... (outros campos serão adicionados mais tarde)
 } task_t ;
